Attempted second teaspoon test
