--- spec/editIdea_spec.js
@@ -1,18 +1,0 @@
-//=require editIdea
-
-describe('updateTitle', function() {
-  it("can alter a post's title", function() {
-    var post = setPost;
-    assert.equal($(post).length, 1))
-  })
-})
-
-function setPost(){
-  var ideas = document.createElement('div');
-  ideas.className = 'ideas';
-  var idea1 = document.createElement('div');
-  idea1.className = 'idea';
-  var title = document.body.idea1 = {
-    title: Hoo;
-  }
-}
