+++ test/test_helper.rb
@@ -4,7 +4,6 @@ require 'rails/test_help'
 require "minitest/rails/capybara"
 require 'factory_girl_rails'
 require 'capybara/rails'
-require 'selenium-webdriver'
 
 class ActiveSupport::TestCase
   # Setup all fixtures in test/fixtures/*.yml for all tests in alphabetical order.
